# skillhub_CICD_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SkillHub API",
        "description": "API REST de la plateforme SkillHub : authentification, formations, modules, inscriptions et messagerie.",
        "contact": {
            "name": "MU202612"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inscrire un utilisateur",
                "description": "Crée un nouveau compte et renvoie un JWT directement utilisable.\n\nÉtapes :\n- Valide les champs requis (nom, email, password, role).\n- Crée le User en base (le password est hashé par le cast Eloquent).\n- Émet un JWT via JWTAuth::fromUser pour que le front enchaîne sans re-login.\n\nUtilise le modèle User et la facade JWTAuth.\nRenvoie un JSON { message, user, token } avec status 201.",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Alice Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "alice@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "MotDePasse123!"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "apprenant",
                                        "enum": [
                                            "apprenant",
                                            "formateur"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé"
                    },
                    "422": {
                        "description": "Champ invalide ou email déjà utilisé"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Se connecter et obtenir un JWT",
                "description": "Vérifie email + password et renvoie un JWT en cas de succès.\n\nÉtapes :\n- Valide email et password.\n- Tente JWTAuth::attempt avec ces credentials.\n- Si l'attempt rate, renvoie 401 sans détail (on ne dit pas ce qui a foiré).\n\nUtilise la facade JWTAuth et auth()->user() pour récupérer le profil.\nRenvoie { message, token, user } ou 401 en cas d'échec.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Email ou mot de passe incorrect"
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Profil de l'utilisateur connecté",
                "description": "Renvoie le profil de l'utilisateur identifié par le token.\n\nÉtapes :\n- Lit le token via authentifierUtilisateur (helper du parent).\n- Si le token est invalide, renvoie l'erreur déjà préparée.\n- Sinon renvoie { user }.",
                "operationId": "d32e56b6274eb68e980da6d91fbe9fd8",
                "responses": {
                    "200": {
                        "description": "Profil renvoyé"
                    },
                    "401": {
                        "description": "Token invalide ou absent"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Invalider le token courant",
                "description": "Invalide le token JWT côté serveur.\n\nLe token est ajouté à la blacklist, donc même si le client le garde\nil ne pourra plus l'utiliser.",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie"
                    },
                    "401": {
                        "description": "Token invalide"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profil/photo": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Uploader une photo de profil",
                "description": "Remplace la photo de profil de l'utilisateur connecté.\n\nÉtapes :\n- Vérifie le JWT.\n- Valide le fichier (image, max 2 Mo, formats jpeg/png/jpg/gif).\n- Supprime l'ancienne photo si elle existe sur disque.\n- Déplace le nouveau fichier vers public/images/profils avec un nom unique\n  (profil_{userId}_{timestamp}.{ext}).\n- Met à jour user.photo_profil et persiste.\n\nUtilise public_path() pour le stockage et $request->file('photo').\nRenvoie { message, photo_profil, user }.",
                "operationId": "446a08dbe9d8b2af53ec54e79c3158cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "photo"
                                ],
                                "properties": {
                                    "photo": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Photo mise à jour"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "422": {
                        "description": "Fichier invalide"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT renvoyé par /api/login. Préfixer avec 'Bearer '.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}